Query cleanup.
